backend/requirements.txt
asgiref==3.11.1
Django==5.2.13
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2026.2
spacy==3.8.14
PyMuPDF==1.25.3
Jinja2==3.1.6
python-dotenv==1.2.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
